added nickname input

+++ client/src/core/Camera.cpp
@@ -11,7 +11,10 @@ SDL_Rect Camera::transformToWorldCordinates(SDL_Rect rect) {
     // flip Y axis
     rect.y = -rect.y;
     // apply scale to texture
-    rect = {rect.x / m_scale, rect.y / m_scale, rect.w / m_scale, rect.h / m_scale };
+    rect = {
+        int(rect.x / m_scale), int(rect.y / m_scale),
+        int(rect.w / m_scale), int(rect.h / m_scale)
+    };
     // move (0,0) to be the center of the screen. Add camera position
     rect.x +=  m_width/2 - m_x/m_scale;
     rect.y +=  m_height/2 + m_y/m_scale;
@@ -21,12 +24,12 @@ SDL_Rect Camera::transformToWorldCordinates(SDL_Rect rect) {
     return rect;
 }
 void Camera::update() {
-  /*
+  // set viewport (x,y) top left corner and (w,h) bottom right
   m_viewport = {
       int(m_x + (m_width * 0.5f * m_scale)), int(m_y + (m_height * 0.5f * m_scale)),
       int(m_x - (m_width * 0.5f * m_scale)), int(m_y - (m_height * 0.5f * m_scale))
   };
-  */
+  // restrct scale
   if(m_scale < 0.01)
     m_scale = 0.01;
 }
